# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Mess Attendance & Billing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessMessage {
  message: string;
}

export interface Resident {
  id: number;
  name: string;
  roomNumber: string;
  whatsappNumber: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateResidentRequest {
  name: string;
  roomNumber: string;
  whatsappNumber: string;
  isActive?: boolean;
}

export type AttendanceRecordStatus =
  (typeof AttendanceRecordStatus)[keyof typeof AttendanceRecordStatus];

export const AttendanceRecordStatus = {
  present: "present",
  half: "half",
  absent: "absent",
  breakfast: "breakfast",
} as const;

export interface AttendanceRecord {
  id: number;
  residentId: number;
  residentName: string;
  date: string;
  status: AttendanceRecordStatus;
  createdAt: string;
}

export type MarkAttendanceRequestStatus =
  (typeof MarkAttendanceRequestStatus)[keyof typeof MarkAttendanceRequestStatus];

export const MarkAttendanceRequestStatus = {
  present: "present",
  half: "half",
  absent: "absent",
  breakfast: "breakfast",
} as const;

export interface MarkAttendanceRequest {
  residentId: number;
  date: string;
  status: MarkAttendanceRequestStatus;
}

export type BulkAttendanceRequestEntriesItemStatus =
  (typeof BulkAttendanceRequestEntriesItemStatus)[keyof typeof BulkAttendanceRequestEntriesItemStatus];

export const BulkAttendanceRequestEntriesItemStatus = {
  present: "present",
  half: "half",
  absent: "absent",
  breakfast: "breakfast",
} as const;

export type BulkAttendanceRequestEntriesItem = {
  residentId: number;
  status: BulkAttendanceRequestEntriesItemStatus;
};

export interface BulkAttendanceRequest {
  date: string;
  entries: BulkAttendanceRequestEntriesItem[];
}

export interface MessSettings {
  id: number;
  messName: string;
  dietRatePerDay: number;
  breakfastRate: number;
  currency: string;
  updatedAt: string;
}

export interface UpdateSettingsRequest {
  messName: string;
  dietRatePerDay: number;
  breakfastRate: number;
  currency: string;
}

export interface ResidentBill {
  residentId: number;
  residentName: string;
  roomNumber: string;
  whatsappNumber: string;
  presentDays: number;
  halfDays: number;
  breakfastDays: number;
  absentDays: number;
  totalDays: number;
  dietRate: number;
  totalAmount: number;
}

export interface BillingSummary {
  month: number;
  year: number;
  messName: string;
  dietRatePerDay: number;
  breakfastRate: number;
  currency: string;
  bills: ResidentBill[];
  totalCollectable: number;
}

export interface SendBillsRequest {
  month: number;
  year: number;
}

export type SendBillsResponseResultsItem = {
  residentId: number;
  residentName: string;
  whatsappNumber: string;
  success: boolean;
  message: string;
};

export interface SendBillsResponse {
  sent: number;
  failed: number;
  results: SendBillsResponseResultsItem[];
}

export interface SendReminderRequest {
  residentId: number;
  month: number;
  year: number;
  customMessage?: string;
}

export type GetAttendanceParams = {
  month?: number;
  year?: number;
  residentId?: number;
  date?: string;
};

export type GetBillingSummaryParams = {
  month: number;
  year: number;
};