# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hostel Mess Attendance & Billing API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: residents
    description: Resident management
  - name: attendance
    description: Attendance tracking
  - name: settings
    description: Mess settings
  - name: billing
    description: Billing and reports
  - name: whatsapp
    description: WhatsApp messaging

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /residents:
    get:
      operationId: getResidents
      tags: [residents]
      summary: Get all residents
      responses:
        "200":
          description: List of residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resident"
    post:
      operationId: createResident
      tags: [residents]
      summary: Add a new resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentRequest"
      responses:
        "201":
          description: Resident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"

  /residents/{id}:
    put:
      operationId: updateResident
      tags: [residents]
      summary: Update a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentRequest"
      responses:
        "200":
          description: Resident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
    delete:
      operationId: deleteResident
      tags: [residents]
      summary: Delete a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resident deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  /attendance:
    get:
      operationId: getAttendance
      tags: [attendance]
      summary: Get attendance records
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: residentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for a resident on a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAttendanceRequest"
      responses:
        "200":
          description: Attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  /attendance/bulk:
    post:
      operationId: bulkMarkAttendance
      tags: [attendance]
      summary: Mark attendance for all residents on a date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAttendanceRequest"
      responses:
        "200":
          description: Attendance marked for all
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get mess settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update mess settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessSettings"

  /billing/summary:
    get:
      operationId: getBillingSummary
      tags: [billing]
      summary: Get billing summary for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Billing summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingSummary"

  /whatsapp/send-bills:
    post:
      operationId: sendMonthlyBills
      tags: [whatsapp]
      summary: Send monthly bills via WhatsApp to all residents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendBillsRequest"
      responses:
        "200":
          description: Bills sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendBillsResponse"

  /whatsapp/send-reminder:
    post:
      operationId: sendReminder
      tags: [whatsapp]
      summary: Send payment reminder to a specific resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendReminderRequest"
      responses:
        "200":
          description: Reminder sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    Resident:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        roomNumber:
          type: string
        whatsappNumber:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - roomNumber
        - whatsappNumber
        - isActive
        - createdAt

    CreateResidentRequest:
      type: object
      properties:
        name:
          type: string
        roomNumber:
          type: string
        whatsappNumber:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - roomNumber
        - whatsappNumber

    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [present, half, absent, breakfast]
        createdAt:
          type: string
      required:
        - id
        - residentId
        - residentName
        - date
        - status
        - createdAt

    MarkAttendanceRequest:
      type: object
      properties:
        residentId:
          type: integer
        date:
          type: string
        status:
          type: string
          enum: [present, half, absent, breakfast]
      required:
        - residentId
        - date
        - status

    BulkAttendanceRequest:
      type: object
      properties:
        date:
          type: string
        entries:
          type: array
          items:
            type: object
            properties:
              residentId:
                type: integer
              status:
                type: string
                enum: [present, half, absent, breakfast]
            required:
              - residentId
              - status
      required:
        - date
        - entries

    MessSettings:
      type: object
      properties:
        id:
          type: integer
        messName:
          type: string
        dietRatePerDay:
          type: number
        breakfastRate:
          type: number
        currency:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - messName
        - dietRatePerDay
        - breakfastRate
        - currency
        - updatedAt

    UpdateSettingsRequest:
      type: object
      properties:
        messName:
          type: string
        dietRatePerDay:
          type: number
        breakfastRate:
          type: number
        currency:
          type: string
      required:
        - messName
        - dietRatePerDay
        - breakfastRate
        - currency

    ResidentBill:
      type: object
      properties:
        residentId:
          type: integer
        residentName:
          type: string
        roomNumber:
          type: string
        whatsappNumber:
          type: string
        presentDays:
          type: number
        halfDays:
          type: number
        breakfastDays:
          type: number
        absentDays:
          type: number
        totalDays:
          type: number
        dietRate:
          type: number
        totalAmount:
          type: number
      required:
        - residentId
        - residentName
        - roomNumber
        - whatsappNumber
        - presentDays
        - halfDays
        - breakfastDays
        - absentDays
        - totalDays
        - dietRate
        - totalAmount

    BillingSummary:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        messName:
          type: string
        dietRatePerDay:
          type: number
        breakfastRate:
          type: number
        currency:
          type: string
        bills:
          type: array
          items:
            $ref: "#/components/schemas/ResidentBill"
        totalCollectable:
          type: number
      required:
        - month
        - year
        - messName
        - dietRatePerDay
        - breakfastRate
        - currency
        - bills
        - totalCollectable

    SendBillsRequest:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
      required:
        - month
        - year

    SendBillsResponse:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              residentId:
                type: integer
              residentName:
                type: string
              whatsappNumber:
                type: string
              success:
                type: boolean
              message:
                type: string
            required:
              - residentId
              - residentName
              - whatsappNumber
              - success
              - message
      required:
        - sent
        - failed
        - results

    SendReminderRequest:
      type: object
      properties:
        residentId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        customMessage:
          type: string
      required:
        - residentId
        - month
        - year