# lib/api-zod/src/generated/types/bulkAttendanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Mess Attendance & Billing API
 * OpenAPI spec version: 0.1.0
 */
import type { BulkAttendanceRequestEntriesItem } from "./bulkAttendanceRequestEntriesItem";

export interface BulkAttendanceRequest {
  date: string;
  entries: BulkAttendanceRequestEntriesItem[];
}