# lib/api-zod/src/generated/types/markAttendanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Mess Attendance & Billing API
 * OpenAPI spec version: 0.1.0
 */
import type { MarkAttendanceRequestStatus } from "./markAttendanceRequestStatus";

export interface MarkAttendanceRequest {
  residentId: number;
  date: string;
  status: MarkAttendanceRequestStatus;
}