# lib/api-zod/src/generated/types/resident.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Mess Attendance & Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface Resident {
  id: number;
  name: string;
  roomNumber: string;
  whatsappNumber: string;
  isActive: boolean;
  createdAt: string;
}