# lib/api-zod/src/generated/types/residentBill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Mess Attendance & Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface ResidentBill {
  residentId: number;
  residentName: string;
  roomNumber: string;
  whatsappNumber: string;
  presentDays: number;
  halfDays: number;
  breakfastDays: number;
  absentDays: number;
  totalDays: number;
  dietRate: number;
  totalAmount: number;
}