# lib/api-zod/src/generated/types/updateSettingsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostel Mess Attendance & Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSettingsRequest {
  messName: string;
  dietRatePerDay: number;
  breakfastRate: number;
  currency: string;
}